Move sklean_helper.* to utils.*
